pyproject.toml
[project]
name = "bsky-umap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gensim==4.3.3",
    "hsluv==5.0.4",
    "numba>=0.61.2",
    "numpy>=1.26.4",
    "pyarrow==19.0.0",
    "python-dotenv==1.0.1",
    "scikit-learn>=1.6.1",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.7",
]
